Change the type of the frame number from uint32_t to int64_t

+++ experimental/Pomdog.Experimental/Gameplay/Scene.hpp
@@ -19,7 +19,7 @@ struct GameScene final {
     std::function<void()> OnExit;
     std::string name;
     ConnectionList connections;
-    Optional<std::uint32_t> frameNumber;
+    Optional<std::int64_t> frameNumber;
 
     void Enter();
